+++ changelog.md
@@ -22,6 +22,7 @@
   * Vigour effect no longer increases stamina regeneration and uses the formula from the enchantment
   * Fixed enchantment formula (-25%/-40%/-45% -> -25%/-40%/-55% consumption)
   * Fixed enchantment not applying to jumps
+* Changed `Bow's Arrows Base Damage` to `Bow's Arrows Base Damage Multiplier`
 * Fixed enchantments info not showing on enchanted books
 * Removed a leftover log